Creating setting for default uploads directory. Also, moving to konduit-serving-api module.
