+++ konduit-serving-api/src/main/java/ai/konduit/serving/settings/constants/Constants.java
